feat: Implement multi-threaded Pi calculator
This commit introduces a new Rust program that calculates the first n
digits of Pi using the BBP algorithm. The program is multi-threaded and
allows the user to specify the number of threads to use. It uses the
`rug` crate for arbitrary-precision arithmetic to ensure the accuracy of
the calculated digits.

The program can be run from the command line, taking the number of
digits and the number of threads as arguments.